Penultimate commit
Added a 60-seconds timer and "Game over"/"Restart" labels (the latter with restarting function), differentiated the pace of all types of duck nodes and the amount of points each type gives, implemented a removal of remaining target nodes as soon as the game is over.

Remaining tasks: updating bullets textures reflecting the current amount of loaded bullets, reloading the weapon, subtracting points when the player misses a valid target and optimization of functions.