It turned out this one should be the penultimate commit
Switched the background color to the one reflecting the user's current system's appearance mode (the app has to be restarted for the effect to take place), set the bullets textures to reflect the current amount of loaded bullets, set all the nuances related to reloading the weapon, and took care of some small cosmetic improvements.

Remaining tasks: subtracting points when the player lets a valid target disappear without landing a shot, optimization of functions.